requirements.txt
einops==0.7.0
ema-pytorch==0.2.3
accelerate==0.21.0
tqdm==4.65.0
wandb==0.16.6
numpy==1.25.1
matplotlib==3.7.2
Pillow==9.5.0
scikit-learn==1.3.2
